Preparato codice per imageview cisterna e pagina accesso wifi esp

+++ app/src/main/java/it/polito/did/did_smartwater/model/Plant.kt
@@ -12,7 +12,7 @@ data class Plant(
     val startTime: String,
     val irrigationDays: Int,
     val humidityLevel: Float,
-    val humidiyThreshold: Float,
+    val humidiyThreshold: Int,
     val note: String
     ) {
     companion object{
